### 2024-09-07 #7193 @editorialbot
[PRE REVIEW]: FluidSF: A Python package for calculating turbulent flow statistics
**Submitting author:** @cassidymwagner (<a href="http://orcid.org/0000-0002-1186-2082">Cassidy Wagner</a>)
**Repository:** https://github.com/cassidymwagner/fluidsf
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
