### 2024-10-24 0ba47ce
updated dois and removed unused references
