+++ paper/paper.bib
@@ -6,8 +6,6 @@
   volume   = {916},
   pages    = {49},
   doi      = {10.1017/jfm.2021.247},
-  keywords = {geostrophic turbulence,rotating turbulence,structure functions,turbulence theory},
-  file     = {/Users/cassswagner/Zotero/storage/UHDFQBS6/full-text.pdf}
 }
 
 @article{wan:2016,
@@ -17,7 +15,8 @@
   volume    = {23},
   number    = {4},
   year      = {2016},
-  publisher = {AIP Publishing}
+  publisher = {AIP Publishing},
+  doi       = {10.1063/1.4945631}
 }
 
 @article{coscarella:2020,
@@ -27,7 +26,8 @@
   volume    = {32},
   number    = {11},
   year      = {2020},
-  publisher = {AIP Publishing}
+  publisher = {AIP Publishing},
+  doi       = {10.1063/5.0028119}
 }
 
 @article{schloesser:2016,
@@ -38,17 +38,8 @@
   number    = {9},
   pages     = {1843--1858},
   year      = {2016},
-  publisher = {American Meteorological Society}
-}
-
-@article{lindborg:2015,
-  title     = {A Helmholtz decomposition of structure functions and spectra calculated from aircraft data},
-  author    = {Lindborg, Erik},
-  journal   = {Journal of Fluid Mechanics},
-  volume    = {762},
-  pages     = {R4},
-  year      = {2015},
-  publisher = {Cambridge University Press}
+  publisher = {American Meteorological Society},
+  doi       = {10.1175/jtech-d-15-0180.1}
 }
 
 @article{balwada2022,
@@ -59,59 +50,8 @@
   number    = {41},
   pages     = {eabq2566},
   year      = {2022},
-  publisher = {American Association for the Advancement of Science}
-}
-
-@article{pearson2020biases,
-  title     = {Biases in structure functions from observations of submesoscale flows},
-  author    = {Pearson, Jenna and Fox-Kemper, Baylor and Pearson, Brodie and Chang, Henry and Haus, Brian K and Horstmann, Jochen and Huntley, Helga S and Kirwan Jr, AD and Lund, Bj{\"o}rn and Poje, Andrew},
-  journal   = {Journal of Geophysical Research: Oceans},
-  volume    = {125},
-  number    = {6},
-  pages     = {e2019JC015769},
-  year      = {2020},
-  publisher = {Wiley Online Library}
-}
-
-@article{pearson2021advective,
-  title     = {Advective structure functions in anisotropic two-dimensional turbulence},
-  author    = {Pearson, Brodie C and Pearson, Jenna L and Fox-Kemper, Baylor},
-  journal   = {Journal of Fluid Mechanics},
-  volume    = {916},
-  pages     = {A49},
-  year      = {2021},
-  publisher = {Cambridge University Press}
-}
-
-@article{xie:2018,
-  title     = {Exact third-order structure functions for two-dimensional turbulence},
-  author    = {Xie, Jin-Han and B{\"u}hler, Oliver},
-  journal   = {Journal of Fluid Mechanics},
-  volume    = {851},
-  pages     = {672--686},
-  year      = {2018},
-  doi       = {10.1017/jfm.2018.528},
-  publisher = {Cambridge University Press}
-}
-
-@inproceedings{kolmogorov:1941,
-  title     = {Dissipation of energy in the locally isotropic turbulence},
-  author    = {Kolmogorov, Andrey Nikolaevich},
-  booktitle = {Doklady Akademii Nauk SSSR},
-  volume    = {32},
-  pages     = {19--21},
-  year      = {1941}
-}
-
-@article{kolmogorov:1991,
[…]
