+++ paper/paper.md
@@ -51,7 +51,7 @@ where $\mathbf{x}$ denotes the position of a data point, $\delta \phi$ denotes t
 
 ## Related Work
 
-`FluidSF` uniquely contributes to the field through a combination of expanded data support, the ability to diagnose a wide array of SF types (including novel and blended SFs), and tools for analyzing spatial variations in SFs. There are several open source software packages available that calculate aspects of spatial SFs. `fastSF` is a parallelized C++ code designed to compute arbitrary-order SFs of velocity or scalars (but not blended) from Cartesian grids of data [@sadhukhan:2021]. @fuchs2022 created an open source `MATLAB` toolkit that performs a variety of turbulence analysis, including arbitrary-order longitudinal-velocity SFs. A complimentary and alternative method to structure functions for analyzing turbulence data is coarse-graining. `FlowSieve` is a primarily C++ package that uses coarse-graining to estimate ocean and atmospheric turbulence properties from Global Climate Model data [@storer2023].
+`FluidSF` uniquely contributes to the field through a combination of expanded data support, the ability to diagnose a wide array of SF types (including novel and blended SFs), and tools for analyzing spatial variations in SFs. `FluidSF` was used to develop new methods for estimating inter-scale geophysical energy fluxes [@pearson:2024]. There are several open source software packages available that calculate aspects of spatial SFs. `fastSF` is a parallelized C++ code designed to compute arbitrary-order SFs of velocity or scalars (but not blended) from Cartesian grids of data [@sadhukhan:2021]. @fuchs2022 created an open source `MATLAB` toolkit that performs a variety of turbulence analysis, including arbitrary-order longitudinal-velocity SFs. A complimentary and alternative method to structure functions for analyzing turbulence data is coarse-graining. `FlowSieve` is a primarily C++ package that uses coarse-graining to estimate ocean and atmospheric turbulence properties from Global Climate Model data [@storer2023].
 
 # Acknowledgements
 
