+++ paper/paper.bib
@@ -183,3 +183,10 @@
   doi       = {10.1038/s41586-020-2649-2},
   publisher = {Springer Science and Business Media {LLC}},
 }
+
+@unpublished{pearson:2024,
+  title     = {A Novel Method to Estimate Spectral Fluxes in Quasi-Two-Dimensional Flows},
+  author    = {Pearson, Brodie and Wagner, Cassidy and {Fox-Kemper}, Baylor and Samelson, Roger},
+  year      = {2024},
+  note      = {Manuscript submitted for publication}
+}
