+++ paper/paper.md
@@ -17,8 +17,8 @@ authors:
     orcid: 0000-0002-0202-0481
     affiliation: '1'
 affiliations:
- - index: 1
  - name: Oregon State University
+   index: 1
 date: 05 August 2024
 bibliography: paper.bib
 ---
