### 2025-03-04 #7193 @philipcardiff
We can continue the conversation over on the review issue: https://github.com/openjournals/joss-reviews/issues/7873.

If other invited reviewers are available, please let me know, and I will add you.
