### 2025-03-12 cassidymwagner/fluidsf#36 @MaceKuailv
Installation and community contribution
This is part of the review for the joss paper https://github.com/openjournals/joss-reviews/issues/7873

There is a couple of things that I would recommend:

- [ ] Submit the package to conda-forge.
- [ ] You should list all dependencies in the documentation and mention which of those are optional. You should also have a environment.yml file for contributors.
- [ ] List the steps of making the pull request. Including how to run the tests and notebooks.
