### 2025-03-12 cassidymwagner/fluidsf#37 @MaceKuailv
Strip notebook output
This is part of the review for the joss paper https://github.com/openjournals/joss-reviews/issues/7873

You should add a couple more steps in .pre-commit-config.yaml (e.g. "black"). In particular, a step that strip the output of notebooks would be very helpful for developers (https://github.com/kynan/nbstripout).
