### 2025-03-12 cassidymwagner/fluidsf#38 @MaceKuailv
Some recommendations on the documentation
This is part of the review for the joss paper https://github.com/openjournals/joss-reviews/issues/7873

I will add items in this issue slowly as I read the documentation more carefully.

- [ ] There are a lot of housekeeping steps in the tutorial notebooks like `import warning` and `ax.set_xlabel`. It would be easier for the reader if you could [collapse the input](https://jupyterbook.org/en/stable/interactive/hiding.html#hide-cell-inputs)
- [ ] Upload software onto zenodo, and add an option to cite the software [here](https://cassidymwagner.github.io/fluidsf/index.html#citing)
- [ ] Also upload your example datasets to zenodo. It takes a lot of navigation to find the data set that you are using for the notebooks, and it is very unlikely to find the exact data you used. You can use something like [pooch](https://pypi.org/project/pooch/) to let the users download the data automatically.
- [ ] It seems like the notebooks in the documentation are not executed during deployment. It could be a good idea if the notebooks in the documentation is updated "on push" in the ci process (https://docs.readthedocs.com/platform/stable/guides/jupyter.html).
