### 2025-03-12 cassidymwagner/fluidsf#37 @MaceKuailv
This needs to happen after resolving the notebook execution issue in #38
