### 2025-03-12 cassidymwagner/fluidsf#39 @MaceKuailv
Get rid of github action on different operating systems
This is part of the review for the joss paper https://github.com/openjournals/joss-reviews/issues/7873

Since this is a package that does not depends on many libraries, I don't think it make a lot of sense to run the tests on different operating systems whenever github action is triggered. Keep the ubuntu, and get rid of the windows and macOS. Those are things people who work on scipy and numpy needs to worry about.

Also, it seems like the tests in the most recent PRs has failed, and the log has expired... For example, PR #34. Please check what is happening.
