### 2025-03-12 cassidymwagner/fluidsf#40 @MaceKuailv
An easy way to boost performance (optional)
This is an optional recommendation as part of the review for the joss paper https://github.com/openjournals/joss-reviews/issues/7873
Feel free to close this if you don't want to implement this.

It seems like most of the function is based on numpy, except for the calculation of great_circle distance using geopy, which means you could compile most of your functions using [numba](https://numba.pydata.org/) decorator. Actually, it is not too hard to rewrite [great_circle](https://github.com/geopy/geopy/blob/master/geopy/distance.py) using just numpy.
