### 2025-03-12 cassidymwagner/fluidsf#41 @MaceKuailv
Higher order gradient calculation
This is part of the JOSS paper review https://github.com/openjournals/joss-reviews/issues/7873

`np.gradient` is used for most of the gradient calculation, which is a second order gradient operator. My worry is that it may not be enough for higher order gradients. It would be great if you can implement some higher order gradient functions. At least, the users should get a warning when doing something like
```
sf = fluidsf.generate_structure_functions_2d(U, V, x, y, sf_type=["LLLLLLLLL"], boundary=None)
```
