### 2025-03-12 cassidymwagner/fluidsf#42 @MaceKuailv
Implement laziness (optional)
This is an optional recommendation as part of my review on the JOSS paper https://github.com/openjournals/joss-reviews/issues/7873

The `sf_dict` is a plain dictionary now. The user needs to know what they need from the very start and trigger all calculation all at once. I am proposing an alternative way of doing this:

```
class sf_dict(dict):
    def __init__(self,x,y,u,v):
        self.u = self.u
        fill_in_the_data()
    def __getitem__(self,item):
        if self.get(item) is None:
            self[item] = self.some_calculation()
        return self.get(item)
   def some_calculation(self):
        return some_function_outside_the_class(self.u,self.x,....)
```
In this way, the calculation is triggered only when you actually need a particular kind of structure function. The next time you want the same structure function there is also no additional cost.
