### 2025-03-19 23af8d7
Define jargon and elaborate on grids
