+++ paper/paper.md
@@ -25,25 +25,25 @@ bibliography: paper.bib
 
 # Summary
 
-Fluid systems are everywhere, from small-scale engineering problems to planetary-and-larger-scale systems (atmosphere, ocean, galactic gas clouds). These systems are often turbulent, where motion is chaotic, unpredictable, and can only be characterized through statistical analyses. Spatial structure functions (SFs) are one such statistical analysis technique for turbulence, that require calculation of spatial differences in properties as a function of their separation distance. By combining and then averaging these spatial differences, various types of SF can be constructed to measure physical properties of fluid flow, such as heat and energy transfers, energy density, intermittency etc. However, calculating SFs is often a cumbersome and computationally-intensive task tailored to the specific format of a given fluid dataset. `FluidSF` is a flexible `Python` package that can be used to diagnose and analyze various physically-informative SFs from 1-, 2-, or 3-dimensional fluid data sets.    
+Fluid systems are everywhere, from small-scale engineering problems to planetary-and-larger-scale systems (atmosphere, ocean, galactic gas clouds). These systems are often turbulent, where motion is chaotic, unpredictable, and can only be characterized through statistical analyses. Structure functions (SFs) are one such statistical analysis technique for turbulence, that require calculation of spatial differences in properties as a function of their separation distance. By combining and then averaging these spatial differences, various types of SF can be constructed to measure physical properties of fluid flow, such as heat and energy transfers, energy density, intermittency etc. However, calculating SFs is often a cumbersome and computationally-intensive task tailored to the specific format of a given fluid dataset. `FluidSF` is a flexible `Python` package that can be used to diagnose and analyze various physically-informative SFs from 1-, 2-, or 3-dimensional fluid data sets.    
 
 # Statement of need
 
-`FluidSF` can construct an array of traditional and modern structure functions (SFs), and can be easily modified to calculate user-defined SFs that utilize general fluid properties, including scalars (e.g., vorticity, density) and/or vectors (e.g., velocity, magnetic field). `FluidSF` also includes several tools to process data (e.g., array shifting, binning) and diagnose useful properties (e.g., advection). The flexibility of this package enables geophysical, astrophysical, and engineering applications such as: quantifying the energy cycles within Earth's ocean [@pearson2019; @balwada2022], Earth's atmosphere [@lindborg:1999], and Jupiter's atmosphere [@young:2017], diagnosing the intermittency of magnetohydrodynamic plasma turbulence [@wan:2016] and the scaling laws of idealized 3D turbulence [@iyer:2020], or measuring the characteristics of ocean surface temperature [@schloesser:2016] or the anistropy of flow over rough beds [@coscarella:2020].   
[…]
