### 2025-03-19 271feda
Update Pearson et al., 2025 references in text
