+++ paper/paper.md
@@ -43,7 +43,7 @@ where $\mathbf{x}$ denotes the position of a data point, $\delta \phi$ denotes t
 
 ![Various structure functions (SFs) calculated from a simulated 2D turbulent flow, visualized through snapshots of the vorticity field (top left) and velocity field (bottom left). The right panels show various SFs based on velocity (red lines) and vorticity (blue lines), including third-order and advective SFs (top right) and traditional second-order SFs (bottom right). The results are from the top layer snapshot of an anisotropic 2-layer quasi-geostrophic simulation conducted with GeophysicalFlows.jl. \label{fig:fig1}](figs/fig1.png)
 
-As demonstrated in \autoref{fig:fig1} and \autoref{fig:fig2}, `FluidSF` can calculate a wide array of traditional structure functions, including $SF_{\phi \phi}$ (\autoref{eq:eq1}; where the scalar field in this case is vorticity $\omega$), second- and third-order SFs of longitudinal velocity ($SF_{LL}=\overline{(\delta u_L)^2}$ and $SF_{LLL}=\overline{(\delta u_L)^3}$; where $u_L=\mathbf{u}\cdot\hat{\mathbf{r}}$) and transverse velocity ($SF_{TT}$ and $SF_{TTT}$), and blended velocity-scalar third-order SFs ($SF_{L\omega\omega}=\overline{\delta u_L \delta \omega \delta \omega}$), in addition to novel advective SFs  of velocity ($ASF_{V}$), vorticity ($ASF_{\omega}$) and scalars [@pearson:2021]. Advective SFs require fields of the local advection, and `FluidSF` has a built-in function to compute these advection terms. `FluidSF` can calculate SFs in specific separation directions (i.e., aligned with the Cartesian co-ordinates, shown in \autoref{fig:fig2}), and for 2D data it can diagnose maps showing how SFs vary with the magnitude and orientation of the separation vector $\mathbf{r}$ (\autoref{fig:fig3}). `FluidSF` also includes tools to make the calculation and processing of SFs easier, such as array shifting, diagnosis of the advection terms for novel SFs, decomposition of velocity into longitudinal (along-$\mathbf{r}$; $u_L$) and transverse (across-$\mathbf{r}$; $u_T$) components, and data binning based on separation distance.
[…]
