### 2025-03-27 d07e351
added restriction to run only when pushing to main
