+++ .github/workflows/publish-to-test-pypi.yml
@@ -1,6 +1,9 @@
 name: Publish Python 🐍 distribution 📦 to PyPI and TestPyPI
 
-on: push
+on: 
+  push:
+    branches:
+      - main
 
 jobs:
   build:
