### 2025-03-27 b6ce656
removing tag input to fix detached head when commiting in GH action
