+++ .github/workflows/draft-pdf.yml
@@ -2,8 +2,8 @@ on:
   push:
     branches:
       - main
-    tags:
-      - "paper*"
+    # tags:
+    #   - "paper*"
   pull_request:
     branches:
       - main
