+++ .github/workflows/draft-pdf.yml
@@ -14,6 +14,10 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
+        with:
+          fetch-depth: 2
+      - run:
+          git checkout HEAD^
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
