### 2025-03-27 f408c14
another attempt to fix detached head
