+++ .github/workflows/draft-pdf.yml
@@ -2,8 +2,8 @@ on:
   push:
     branches:
       - main
-    # tags:
-    #   - "paper*"
+    tags:
+      - "paper*"
   pull_request:
     branches:
       - main
@@ -14,10 +14,6 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
-        with:
-          fetch-depth: 2
-      - run:
-          git checkout HEAD^
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
@@ -37,3 +33,4 @@ jobs:
         with:
           message: '(auto) Paper PDF Draft'
           add: 'paper/paper.pdf'
+          push: origin HEAD:{{ github.head_ref }}
