### 2025-03-27 9764a82
if statement for checkout to work better with PRs
