+++ .github/workflows/draft-pdf.yml
@@ -14,6 +14,12 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
+      - name: Checkout for PRs
+        if: ${{ github.event_name == 'pull_request' }}
+        uses: actions/checkout@v4
+        with:
+          repository: ${{ github.event.pull_request.head.repo.full_name }}
+          ref: ${{ github.event.pull_request.head.ref }}
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
@@ -33,4 +39,3 @@ jobs:
         with:
           message: '(auto) Paper PDF Draft'
           add: 'paper/paper.pdf'
-          push: origin HEAD:${{ github.head_ref }}
