### 2025-03-27 2898706
merge some commits from 39-get-rid-of-github-action-on-different-operating-systems
