### 2025-03-27 cassidymwagner/fluidsf#38 @cassidymwagner
Thank you for these comments! We have been unsure of how to handle example datasets because we have two types: simulated data generated with open source numerical modeling packages and NASA/AVISO datasets that require login information to access.

We could easily upload the first type of dataset, but the NASA/AVISO datasets should be accessed only through their appropriate channels. There are a few Python packages for easily downloading NASA data (some okay options for AVISO) that could be included in a notebook, but those still require an email and password. Do you have any suggestions for how to provide sufficient access to this kind of data?

This problem is why the notebooks are not executed during deployment but instead locally, where I have access to the NASA/AVISO data.
