### 2025-03-28 cassidymwagner/fluidsf#38 @MaceKuailv
OK, if there are rules against that, then we shouldn't upload them. Sorry, I didn't know about this.

There are two options I think:
1. leave this particular notebook as it is and exclude it from executing.
2. create some synthetic data in the same format. You could either regrid eta in some simulation or just do something like cos(lat)*cos(lon).

I personally prefer option 2 because you can build a sense of trust when your user click "restart and run all" and everything went smoothly.
