### 2025-04-08 cassidymwagner/fluidsf#45 @ManuelRatz
List the optional dependencies and add additional options
This is part of the review for the joss paper https://github.com/openjournals/joss-reviews/issues/7873

The examples require some optional packages (e.g. seaborn, matplotlib) which are not installed automatically.

It would be nice if these could be installed as optional dependencies listed in the pyproject.toml file using another keyword such as 'examples'.

Also, it would be nice if these optional dependencies are also listed on the github mainpage. Speaking from experience with colleagues, people typically check the mainpage for which packages are required.
