### 2025-04-08 cassidymwagner/fluidsf#46 @ManuelRatz
Include .py files of the examples (optional)
This is part of the review for the joss paper https://github.com/openjournals/joss-reviews/issues/7873

The examples are given as jupyter notebooks which are nice for instructions. Out of convenience, it would be nice if these could also be given as plain .py files with comments that can be more easily accessed/debugged by other IDEs.
