### 2025-04-08 cassidymwagner/fluidsf#47 @ManuelRatz
Make data easier to locate
This is part of the review for the joss paper https://github.com/openjournals/joss-reviews/issues/7873

This is in line with #38 since finding the data for the tutorials can be hard. It would be nice if these are either included directly or can be downloaded from another server. This helps to better debug/understand the different tutorials which can be hard at the moment.
