### 2025-04-08 cassidymwagner/fluidsf#48 @ManuelRatz
Formatting of the returned variables in the API
This is part of the review for the joss paper https://github.com/openjournals/joss-reviews/issues/7873

The formatting in the API return of the functions returning dicts are very messsy. I recommend to use dashes or another form of subindexing. Also, for the claculate_* functions, the full dict is explained while for the generate_* functions (which are used by the user), they are not.
