### 2025-04-08 cassidymwagner/fluidsf#49 @ManuelRatz
Remove Nones from the return dictionaries
This is part of the review for the joss paper https://github.com/openjournals/joss-reviews/issues/7873

The return dictionary always includes the full set of all possible structure functions, even if many of these are empty. For beginners, it might be better to strip the None values and only keep the dictionaries that actually contain data.
