### 2025-04-08 #7873 @ManuelRatz
Upfront: My apologies for the late review. I was swamped with a deadline that is now gone.

I finished my review with some (mostly) small remarks that can be easily solved. I will tick the remaining boxes when the issues are closed. Overall, I think the package is very well documented and the paper is well written.

Please let me know if some comments are unclear.
