### 2025-04-16 7bf77d0
add optional dependencies to run examples
