+++ pyproject.toml
@@ -14,10 +14,11 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
 ]
-dependencies = ["numpy>=1,<3", "pandas>=1.5.3", "geopy>=2"]
+dependencies = ["numpy>=1,<3", "pandas>=2.1.0", "geopy>=2"]
 
 [project.optional-dependencies]
 test = ["pytest>=7,<8"]
+examples = ["matplotlib>=3.7.0", "seaborn>=0.12.0", "h5py>=3.10.0", "scipy>=1.11.0", "xarray>=2024.9.0"]
 
 [project.urls]
 "Homepage" = "https://github.com/cassidymwagner/fluidsf"
