### 2025-04-16 781c237
Update citation to our paper under revision
