+++ paper/paper.bib
@@ -185,8 +185,8 @@
 }
 
 @unpublished{pearson:2025,
-  title     = {A Novel Method to Estimate Spectral Fluxes in Quasi-Two-Dimensional Flows},
+  title     = {Estimating Spectral Fluxes in Quasi-Two-Dimensional Flows with Advective Structure Functions and Bessel Functions},
   author    = {Pearson, Brodie and Wagner, Cassidy and {Fox-Kemper}, Baylor and Samelson, Roger},
-  year      = {2025},
+  year      = {Under Revision},
   note      = {Under Revision in Journal of Physical Oceanography}
 }
