### 2025-04-16 3ad7a42
strip Nones from returned dictionary
