### 2025-04-22 c62259a
strip Nones from returned dict
