### 2025-04-24 2665b25
docstring note  that sf_type is a list of strings
