### 2025-04-24 38e9f32
reformatted return docstring to be clearer with sphinx docs
