### 2025-04-24 993ea9e
moved dict keys back to returns and formatted nicely for sphinx docs
