### 2025-04-24 b599ab6
docstring note that sf_type list entries must be string
