### 2025-04-24 59b1032
reformatted dict values in docstring
