+++ src/fluidsf/calculate_structure_function_3d.py
@@ -61,37 +61,62 @@ def calculate_structure_function_3d(  # noqa: D417, C901
         dict:
             A dictionary containing the velocity-based advective structure functions and
             scalar structure functions (if applicable).
-            The dictionary has the following keys:
-                'SF_velocity_x': The velocity-based advective structure function in the
+            The returned dictionary may contain the following keys, with some keys
+            removed if the structure function is not calculated:
+
+                **SF_velocity_x**: The velocity-based advective structure function in the
                 x direction.
-                'SF_velocity_y': The velocity-based advective structure function in the
+
+                **SF_velocity_y**: The velocity-based advective structure function in the
                 y direction.
-                'SF_velocity_z': The velocity-based advective structure function in the
+
+                **SF_velocity_z**: The velocity-based advective structure function in the
                 z direction.
-                'SF_scalar_x': The scalar-based advective structure function in the x
+
+                **SF_scalar_x**: The scalar-based advective structure function in the x
                 direction.
-                'SF_scalar_y': The scalar-based advective structure function in the y
+
+                **SF_scalar_y**: The scalar-based advective structure function in the y
                 direction.
-                'SF_scalar_z': The scalar-based advective structure function in the z
+
+                **SF_scalar_z**: The scalar-based advective structure function in the z
                 direction.
-                'SF_LL_x': The traditional structure function LL in the x direction.
-                'SF_LL_y': The traditional structure function LL in the y direction.
-                'SF_LL_z': The traditional structure function LL in the z direction.
-                'SF_TT_x': The traditional structure function TT in the x direction.
-                'SF_TT_y': The traditional structure function TT in the y direction.
-                'SF_TT_z': The traditional structure function TT in the z direction.
-                'SF_SS_x': The traditional structure function SS in the x direction.
-                'SF_SS_y': The traditional structure function SS in the y direction.
-                'SF_SS_z': The traditional structure function SS in the z direction.
-                'SF_LLL_x': The traditional structure function LLL in the x direction.
-                'SF_LLL_y': The traditional structure function LLL in the y direction.
-                'SF_LLL_z': The traditional structure function LLL in the z direction.
-                'SF_LTT_x': The traditional structure function LTT in the x direction.
-                'SF_LTT_y': The traditional structure function LTT in the y direction.
-                'SF_LTT_z': The traditional structure function LTT in the z direction.
-                'SF_LSS_x': The traditional structure function LSS in the x direction.
-                'SF_LSS_y': The traditional structure function LSS in the y direction.
-                'SF_LSS_z': The traditional structure function LSS in the z direction.
+
+                **SF_LL_x**: The traditional structure function LL in the x direction.
+
+                **SF_LL_y**: The traditional structure function LL in the y direction.
+
+                **SF_LL_z**: The traditional structure function LL in the z direction.
+
+                **SF_TT_x**: The traditional structure function TT in the x direction.
+
+                **SF_TT_y**: The traditional structure function TT in the y direction.
+                
+                **SF_TT_z**: The traditional structure function TT in the z direction.
+                
+                **SF_SS_x**: The traditional structure function SS in the x direction.
+
+                **SF_SS_y**: The traditional structure function SS in the y direction.
+
+                **SF_SS_z**: The traditional structure function SS in the z direction.
+
+                **SF_LLL_x**: The traditional structure function LLL in the x direction.
+
+                **SF_LLL_y**: The traditional structure function LLL in the y direction.
+
+                **SF_LLL_z**: The traditional structure function LLL in the z direction.
+
+                **SF_LTT_x**: The traditional structure function LTT in the x direction.
+
[…]
