### 2025-04-24 60d14ee
note in docstring for string type in sf_type
