+++ src/fluidsf/calculate_sf_maps_2d.py
@@ -41,7 +41,8 @@ def calculate_sf_maps_2d(  # noqa: D417, C901
             Shift amount for y shift.
         sf_type: list
             List of structure function types to calculate.
-            Accepted types are: "ASF_V", "ASF_S", "LL", "TT", "SS", "LLL", "LTT", "LSS".
+            Accepted list entries must be one or more of the following strings:
+            "ASF_V", "ASF_S", "LL", "TT", "SS", "LLL", "LTT", "LSS".
         scalar: ndarray, optional
             Array of scalar values. Defaults to None.
         adv_scalar: ndarray, optional
