### 2025-04-24 4f5b3fb
formatted docstring dict section
