+++ src/fluidsf/generate_structure_functions_1d.py
@@ -176,14 +176,14 @@ def generate_structure_functions_1d(  # noqa: C901, D417
     data = {
         key: value
         for key, value in {
-        "SF_LL": SF_LL,
-        "SF_TT": SF_TT,
-        "SF_SS": SF_SS,
-        "SF_LLL": SF_LLL,
-        "SF_LTT": SF_LTT,
-        "SF_LSS": SF_LSS,
-        "x-diffs": xd,
-        }.items()
+            "SF_LL": SF_LL,
+            "SF_TT": SF_TT,
+            "SF_SS": SF_SS,
+            "SF_LLL": SF_LLL,
+            "SF_LTT": SF_LTT,
+            "SF_LSS": SF_LSS,
+            "x-diffs": xd,
+            }.items()
         if value is not None
     }
     return data
