### 2025-04-24 1ef5b4d
stripped nones from returned dict
