### 2025-04-24 86f339f
note that sf_type list must have string entries
