+++ src/fluidsf/generate_structure_functions_3d.py
@@ -40,9 +40,10 @@ def generate_structure_functions_3d(  # noqa: C901, D417
         z: ndarray
             1D array of z-coordinates.
         sf_type: list
-            List of structure function types to calculate. Accepted types are:
-            "ASF_V", "ASF_S", "LL", "TT", "SS", "LLL", "LTT", "LSS". Defaults to
-            ["ASF_V"].
+            List of structure function types to calculate.
+            Accepted list entries must be one or more of the following strings:
+            "ASF_V, "ASF_S", "LL", "TT", "SS", "LLL", "LTT", "LSS".
+            Defaults to ["ASF_V"].
         scalar: ndarray, optional
             3D array of scalar values. Defaults to None.
         boundary: str, optional
