### 2025-04-24 adf036e
added returned dictionary definitions in docstring
