+++ src/fluidsf/generate_structure_functions_3d.py
@@ -58,6 +58,86 @@ def generate_structure_functions_3d(  # noqa: C901, D417
         dict:
             Dictionary containing the requested structure functions and separation
             distances for the x, y, and z directions.
+            The returned dictionary may contain the following keys, with some keys
+            removed if the structure function is not calculated:
+
+                **SF_advection_velocity_x**: The advection velocity structure function
+                in the x direction.
+
+                **SF_advection_velocity_y**: The advection velocity structure function
+                in the y direction.
+
+                **SF_advection_velocity_z**: The advection velocity structure function
+                in the z direction.
+
+                **SF_advection_scalar_x**: The scalar advective structure function
+                in the x direction.
+
+                **SF_advection_scalar_y**: The scalar advective structure function
+                in the y direction.
+
+                **SF_advection_scalar_z**: The scalar advective structure function
+                in the z direction.
+
+                **SF_LL_x**: The second-order longitudinal velocity structure function
+                in the x direction.
+
+                **SF_LL_y**: The second-order longitudinal velocity structure function
+                in the y direction.
+
+                **SF_LL_z**: The second-order longitudinal velocity structure function
+                in the z direction.
+
+                **SF_TT_x**: The second-order transverse velocity structure function
+                in the x direction.
+
+                **SF_TT_y**: The second-order transverse velocity structure function
+                in the y direction.
+
+                **SF_TT_z**: The second-order transverse velocity structure function
+                in the z direction.
+
+                **SF_SS_x**: The second-order scalar structure function in the x
+                direction.
+
+                **SF_SS_y**: The second-order scalar structure function in the y
+                direction.
+
+                **SF_SS_z**: The second-order scalar structure function in the z
+                direction.
+
+                **SF_LLL_x**: The third-order longitudinal velocity structure function
+                in the x direction.
+
+                **SF_LLL_y**: The third-order longitudinal velocity structure function
+                in the y direction.
+
+                **SF_LLL_z**: The third-order longitudinal velocity structure function
+                in the z direction.
+
+                **SF_LTT_x**: The longitudinal-transverse velocity structure function
+                in the x direction.
+
+                **SF_LTT_y**: The longitudinal-transverse velocity structure function
+                in the y direction.
+
+                **SF_LTT_z**: The longitudinal-transverse velocity structure function
+                in the z direction.
+
+                **SF_LSS_x**: The longitudinal-scalar structure function in the x
+                direction.
+
+                **SF_LSS_y**: The longitudinal-scalar structure function in the y
+                direction.
+
+                **SF_LSS_z**: The longitudinal-scalar structure function in the z
+                direction.
+
+                **x-diffs**: The separation distances in the x direction.
+
+                **y-diffs**: The separation distances in the y direction.
+
+                **z-diffs**: The separation distances in the z direction.
 
     """
     # Initialize variables as NoneType
