### 2025-04-24 cassidymwagner/fluidsf#40 @cassidymwagner
Thank you for this suggestion! We are exploring a future shift of FluidSF from numpy to xarray/dask. We have seen some really nice performance boosts in the early stages of this process, not to mention the convenience of using xarray datasets. It looks like [Numba can be easily used with xarray](https://tutorial.xarray.dev/advanced/apply_ufunc/numba-vectorization.html), so we will keep this in mind for the next version of FluidSF!
