### 2025-04-24 cassidymwagner/fluidsf#40 @MaceKuailv
Wow, I didn't know about this development with dask and numba! Glad you are thinking about the next step already.

I will keep this issue open so that I can be notified by, but this won't hinder your publishing of the joss paper.
