### 2025-05-07 3b11482
Add contribution/development info to README.md
