### 2025-05-07 551a429
Update generate_structure_functions_2d.py
