### 2025-05-08 9d02afb
Update generate_structure_functions_2d.py
Minor warning text adjustment
