+++ src/fluidsf/generate_structure_functions_2d.py
@@ -217,7 +217,7 @@ def generate_structure_functions_2d(  # noqa: C901, D417
         if any("LSS" in t for t in sf_type):
             SF_x_LSS[x_shift] = SF_dicts["SF_LSS_x"]
         if any("LLLL" in t for t in sf_type):
-            warnings.warn('Structure functions of order 4+ require manual coding. LL and LLL will still be computed.')
+            warnings.warn('Structure functions of order 4 or higher require manual coding. LL and LLL will still be computed.')
 
         # Calculate separation distances in x
         xd[x_shift], tmp = calculate_separation_distances(
