### 2025-05-08 f024594
Update generate_structure_functions_2d.py
Moved warning location outside of loop
