### 2025-05-08 4a79d4d
Update generate_structure_functions_1d.py
