### 2025-05-08 4c4a3cf
Update generate_structure_functions_3d.py
