### 2025-05-08 7be5d17
Update generate_structure_functions_3d.py
