+++ src/fluidsf/generate_structure_functions_3d.py
@@ -1,4 +1,5 @@
 import numpy as np
+import warnings
 
 from .bin_data import bin_data
 from .calculate_advection_3d import calculate_advection_3d
