### 2025-05-15 2e334dc
created .py versions of jupyter notebooks
