+++ examples/python_scripts/ex_cascade.py
@@ -1,0 +1,207 @@
+"""
+Estimate cascade rates from structure functions.
+
+This example will demonstrate how to estimate cascade rates from two different 
+velocity-based structure functions. Cascade rate estimations depend on the physical 
+constraints of the turbulent system and this example specifically applies to 
+2D quasi-geostrophic fluid dynamics. 
+
+General procedure:
+1. Set up plot environment & load/format the data
+2. Calculate structure functions for all snapshots of the simulation
+3. Bootstrap to generate a mean
+4. Estimate cascade rates
+5. Plot cascade rates as a function of length scale
+"""
+
+
+import warnings
+
+import fluidsf
+import h5py
+import matplotlib.pyplot as plt
+import numpy as np
+import pooch
+from scipy.stats import bootstrap
+
+# Ignore warnings for the purpose of this tutorial
+warnings.filterwarnings("ignore")
+
+# Ignore linting for variable names
+# ruff: noqa: N816
+
+# We will use Pooch to download a sample dataset.
+file_path = pooch.retrieve(
+    url="https://zenodo.org/records/15278227/files/2layer_128.jld2",
+    known_hash="a04abc602ca3bbc4ff9a868a96848b6815a17f697202fb12e3ff40762de92ec6",
+)
+
+# We will use h5py to load a .jld2 file, the output from GeophysicalFlows.jl, a
+# numerical ocean simulator written in Julia. The data consists of 2D (horizontal)
+# fields simulated over a periodic domain. There are multiple snapshots of this data,
+# corresponding to different times. We will select the last snapshot to analyze.
+f = h5py.File(file_path, "r")
+grid = f["grid"]
+snapshots = f["snapshots"]
+
+# Initialize the grid of x and y coordinates
+x = grid["x"][()]
+y = grid["y"][()]
+
+# Grab u and v for all snapshots and layers (we will just use the top layer)
+u = snapshots["u"]
+v = snapshots["v"]
+
+# Make a couple of quick plots of the top layer at the last timestep to see the
+# velocity fields.
+fig, (ax1, ax2) = plt.subplots(1, 2, sharex=True, sharey=True, figsize=(12, 5))
+p1 = ax1.pcolormesh(x, y, u["20050"][0], cmap="RdBu", vmin=-5, vmax=5)
+p2 = ax2.pcolormesh(x, y, v["20050"][0], cmap="RdBu", vmin=-5, vmax=5)
+fig.colorbar(p1, label="u-velocity")
+fig.colorbar(p2, label="v-velocity")
+ax1.set_xlabel("x")
+ax2.set_xlabel("x")
+ax1.set_ylabel("y")
+ax2.set_ylabel("y")
+plt.tight_layout()
+
+# Calculate structure functions for all snapshots
+# Here we will calculate the advective velocity structure function and the third order
+# longitudinal velocity structure function. We will calculate each structure function
+# in the x-direction and the y-direction for a total of four unique structure functions.
+sfs_list = [
+    fluidsf.generate_structure_functions_2d(
+        u[d][0],
+        v[d][0],
+        x,
+        y,
+        sf_type=["ASF_V", "LLL"],
+        boundary="periodic-all",
+    )
+    for d in u.keys()
+]
+
+# Bootstrap structure functions
[…]
