### 2025-05-16 e4ad6da
removed plt.show for CI
