+++ examples/python_scripts/qs.py
@@ -120,7 +120,6 @@ plt.hlines(0, 0, 1, color="k", lw=1, zorder=0)
 ax.set_xlabel("Separation distance (r)")
 ax.set_ylabel("Structure function (SF(r))")
 ax.legend(loc="lower right")
-plt.show()
 
 # Scalar structure functions
 # The scalar structure functions can be generated by including "ASF_S" and/or "LSS"
@@ -170,4 +169,3 @@ plt.hlines(0, 0, 1, color="k", lw=1, zorder=0)
 ax.set_xlabel("Separation distance (r)")
 ax.set_ylabel("Structure function (SF(r))")
 ax.legend(loc="lower right")
-plt.show()
