### 2025-05-16 4b9c1c5
draft CI for running examples on push/pull request
