### 2025-05-16 e6edc9e
typo to fix CI dependencies install
