### 2025-05-16 960f393
added pooch to examples dependencies
