+++ pyproject.toml
@@ -18,7 +18,7 @@ dependencies = ["numpy>=1,<3", "pandas>=2.1.0", "geopy>=2"]
 
 [project.optional-dependencies]
 test = ["pytest>=7,<8"]
-examples = ["matplotlib>=3.7.0", "seaborn>=0.12.0", "h5py>=3.10.0", "scipy>=1.11.0", "xarray>=2024.9.0"]
+examples = ["matplotlib>=3.7.0", "seaborn>=0.12.0", "h5py>=3.10.0", "scipy>=1.11.0", "xarray>=2024.9.0", "pooch>=1.8.0"]
 
 [project.urls]
 "Homepage" = "https://github.com/cassidymwagner/fluidsf"
