### 2025-05-16 2318be6
added xarray backend to examples dependencies
