### 2025-05-20 a14a39e
Update generate_sf_maps_2d.py
Edited some 2D map descriptions
