### 2025-05-20 37ef39a
Update generate_structure_functions_1d.py
