### 2025-05-20 b53cf8f
Update generate_structure_functions_2d.py
Minor edits to variable descriptions
