### 2025-05-20 c775619
Update generate_sf_maps_2d.py
