### 2025-05-20 395cbe4
Update calculate_structure_function_1d.py
