### 2025-05-20 0dd5a02
Update generate_structure_functions_2d.py
