### 2025-05-20 c1406ce
Update generate_structure_functions_3d.py
