### 2025-05-20 50bd52d
Update calculate_sf_maps_2d.py
