+++ src/fluidsf/calculate_sf_maps_2d.py
@@ -17,7 +17,7 @@ def calculate_sf_maps_2d(  # noqa: D417, C901
     adv_scalar=None,
 ):
     """
-    Calculate structure function, either advective or traditional.
+    Calculate structure functions, including advective structure functions.
     Supports velocity-based structure functions and scalar-based structure functions.
     Can only be used with periodic, evenly-spaced data.
 
@@ -56,29 +56,29 @@ def calculate_sf_maps_2d(  # noqa: D417, C901
             The returned dictionary may contain the following keys, with some keys
             removed if the structure function is not calculated:
 
-                **SF_advection_velocity_xy**: The advection velocity structure function
+                **SF_advection_velocity_xy**: The advective velocity structure function
                 for separation vectors in the x-y plane.
 
-                **SF_advection_scalar_xy**: The scalar advective structure function
+                **SF_advection_scalar_xy**: The advective scalar structure function
                 for separation vectors in the x-y plane.
 
-                **SF_LL_xy**: The traditional structure function LL for separation
-                vectors in the x-y plane.
+                **SF_LL_xy**: The second-order longitudinal velocity structure function for 
+                separation vectors in the x-y plane.
 
-                **SF_TT_xy**: The traditional structure function TT for separation
-                vectors in the x-y plane.
+                **SF_TT_xy**: The second-order transverse velocity structure function for 
+                separation vectors in the x-y plane.
 
-                **SF_SS_xy**: The traditional structure function SS for separation
-                vectors in the x-y plane.
+                **SF_SS_xy**: The second-order scalar structure function for 
+                separation vectors in the x-y plane.
 
-                **SF_LLL_xy**: The traditional structure function LLL for separation
-                vectors in the x-y plane.
+                **SF_LLL_xy**: The third-order longitudinal velocity structure function for 
+                separation vectors in the x-y plane.
 
-                **SF_LTT_xy**: The traditional structure function LTT for separation
-                vectors in the x-y plane.
+                **SF_LTT_xy**: The third-order longitudinal-transverse-transverse velocity 
+                structure function for separation vectors in the x-y plane.
 
-                **SF_LSS_xy**: The traditional structure function LSS for separation
-                vectors in the x-y plane.
+                **SF_LSS_xy**: The third-order longitudinal-scalar-scalar structure function 
+                for separation vectors in the x-y plane.
 
     """
     inputs = {
