### 2025-05-20 52ef256
Update calculate_structure_function_2d.py
