### 2025-05-20 3005af7
Update calculate_structure_function_3d.py
Homogenizing names across calculate and generate functions
