### 2025-06-03 cassidymwagner/fluidsf#41 @BrodiePearson
@MaceKuailv Following your comment, we decided to add warnings to higher-order structure functions noting that these functions should be  coded by the user to ensure appropriate discretization.

The gradient operators only appear when computing one type of structure function (the "advective" type which depends on derivatives of fields). This type of structure function does not have a high-order equivalent structure function. However, your suggestion of enabling higher-order gradient computations for increased accuracy is valid. We plan on switching away from numpy in the near future (see response in #40), and we will try to enable user-specified order for gradient computations at that time (with the default being the current second-order gradient).
