+++ src/fluidsf/generate_sf_maps_2d.py
@@ -35,7 +35,7 @@ def generate_sf_maps_2d(  # noqa: C901, D417
         y: ndarray
             1D array of y-coordinates.
         sf_type: list
-            List of structure function types to calculate. Accepted list entries must 
+            List of structure function types to calculate. Accepted list entries must
             be one or more of the following strings:
             "ASF_V", "ASF_S", "LL", "TT", "SS", "LLL", "LTT", "LSS". Defaults to
             ["ASF_V"].
@@ -74,11 +74,11 @@ def generate_sf_maps_2d(  # noqa: C901, D417
                 **SF_LLL_xy**: The third-order longitudinal velocity structure function
                 for separation vectors in the x-y plane.
 
-                **SF_LTT_xy**: The third-order longitudinal-transverse-transverse velocity
-                structure function for separation vectors in the x-y plane.
+                **SF_LTT_xy**: The third-order longitudinal-transverse-transverse
+                velocity structure function for separation vectors in the x-y plane.
 
-                **SF_LSS_xy**: The third-order longitudinal-scalar-scalar structure function
-                for separation vectors in the x-y plane.
+                **SF_LSS_xy**: The third-order longitudinal-scalar-scalar structure
+                function for separation vectors in the x-y plane.
 
                 **separation_distances**: 2D array of separation distances between
                 points in the x-y plane.
