+++ .github/workflows/ci.yml
@@ -44,7 +44,7 @@ jobs:
         coverage run -m pytest tests/
 
     - name: Upload coverage reports to Codecov
-      uses: codecov/codecov-action@v4.0.1
+      uses: codecov/codecov-action@v5.4.3
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
 
