### 2025-06-04 9504064
linting and added stacklevel=2 to warnings to give context
