+++ src/fluidsf/generate_structure_functions_3d.py
@@ -1,5 +1,6 @@
-import numpy as np
 import warnings
+
+import numpy as np
 
 from .bin_data import bin_data
 from .calculate_advection_3d import calculate_advection_3d
@@ -149,7 +150,11 @@ def generate_structure_functions_3d(  # noqa: C901, D417
         SF_y_LSS = np.zeros(len(sep_y) + 1)
         SF_z_LSS = np.zeros(len(sep_z) + 1)
     if any("LLLL" in t for t in sf_type):
-            warnings.warn('Structure functions of order 4 or higher require manual coding. LL and LLL will still be computed.')
+        warnings.warn(
+            "Structure functions of order 4 or higher require manual coding. "
+            "LL and LLL will still be computed.",
+            stacklevel=2,
+        )
 
     # Iterate over separations in x, y, and z
     for x_shift in sep_x:
