### 2025-06-11 cassidymwagner/fluidsf#38 @cassidymwagner
I'm adding a quick summary of the current progress on this issue since there are a few things here I think we have addressed through recent PRs!

- [ ] **Housekeeping steps in tutorial notebooks**
        TODO
- [x] **Upload FluidSF onto Zenodo**
         FluidSF has a [Zenodo link](https://zenodo.org/records/11406185) which is listed as a badge on the README
- [ ] **Add an option to cite the software in citation section of docs**
          TODO, also add to README
- [x] **Execute notebooks during deployment**
         We copied the example notebooks to a top-level examples directory, separate from the documentation directories, and we now execute them (and .py versions) during CI. This approach was a little more straightforward than executing notebooks during Sphinx documentation deployment, and also makes it easier for users/contributors to find the example notebooks and sample data (also living at the top-level of the repo). We will continue to look for ways to streamline this process in future version releases.
- [x] **Upload example datasets to Zenodo**
         We uploaded the simulated datasets and now use Pooch to access them in the example notebooks.
- [x] **Handling NASA datasets**
         We separated the example notebooks that depend on NASA datasets into `examples/advanced_examples`, which are not executed during CI since we don't have a way to provide access to the data through Zenodo/Pooch. [Oceananigans](https://github.com/CliMA/Oceananigans.jl) has directories for [examples](https://github.com/CliMA/Oceananigans.jl/tree/main/examples) (executed during CI) and [validation](https://github.com/CliMA/Oceananigans.jl/tree/main/validation) (not executed or tested but useful resources for setting up complex experiments), so we are adopting a similar strategy.
