### 2025-06-11 c8a0741
added older xarray dependency for compatibility
