### 2025-06-11 7661ad7
reformatting and rearranging, pytest section, optional dependencies, zenodo citation,
