+++ README.md
@@ -5,7 +5,7 @@
 [![codecov](https://codecov.io/github/cassidymwagner/fluidsf/graph/badge.svg?token=1ZZ2HUONX4)](https://codecov.io/github/cassidymwagner/fluidsf)
 
 
-# Overview
+### Overview
 FluidSF is a Python package for calculating structure functions from fluid data. These structure functions can be used to estimate turbulence cascade rates without the constraints of spectral methods. This package serves as a useful tool for analyzing turbulent dynamics in the ocean, atmosphere, and beyond.
 
 <p align="center">
@@ -16,26 +16,21 @@ FluidSF is a Python package for calculating structure functions from fluid data.
 
 **For detailed documentation and examples, see the [FluidSF website](https://cassidymwagner.github.io/fluidsf).**
 
-Installation
----
+### Installation
 The easiest method to install FluidSF is with [pip](https://pip.pypa.io/):
 
-```console
-$ pip install fluidsf
-```
+    $ pip install fluidsf
 
-You can also fork/clone this repository to your local machine and install it locally with pip as well:
+You can also fork/clone this repository to your local machine and install it locally with pip:
 
-```console
-$ pip install .
-```
+    $ pip install .
 
-**Optional dependencies to run example notebooks**: [matplotlib](https://matplotlib.org), [seaborn](https://seaborn.pydata.org), [h5py](https://www.h5py.org), [scipy](https://scipy.org), [xarray](https://xarray.dev)
+**List of optional dependencies to run example notebooks**: [matplotlib](https://matplotlib.org), [seaborn](https://seaborn.pydata.org), [h5py](https://www.h5py.org), [scipy](https://scipy.org), [xarray](https://xarray.dev)
 
 
 
-Quickstart
----
+### Quickstart
+
 Once FluidSF is installed, you can load the module into Python and run some basic calculations with any data. Here we'll initialize linearly increasing velocity fields. For more detail on this example, [see the full notebook on the FluidSF website.](https://cassidymwagner.github.io/fluidsf/qs.html)
 
 First we'll initialize a random 2-D field to analyze:
@@ -73,16 +68,24 @@ plt.show()
 <img src="https://github.com/cassidymwagner/fluidsf/blob/main/docs/quickstart.png" alt="Advective structure function plots" width="400"/>
 </p>
 
-"Can I use FluidSF with my data?"
----
+### "Can I use FluidSF with my data?"
+
 Hopefully! FluidSF was initially developed for numerical simulations and satellite data, but there are of course many different types of data. If you are interested in using this package but you are unsure how to use it with your dataset, please reach out and we are happy to assist! 
 
 The best way to communicate about your data needs is to [start a discussion](https://github.com/cassidymwagner/fluidsf/discussions) where you can describe your dataset and what you're hoping to learn with FluidSF. Before starting a discussion you can check through other discussion posts or review the open (and closed) issues to see if any other users have a similar question or dataset. 
 
 We have plans to support many different types of data, especially oceanographic data, but we encourage any users to engage with us so we can make sure we support as many datasets as possible!
+### Citing
 
-Contributing/Developing Guidelines
----
+If you use FluidSF in your research or educational activities, we would be grateful if you credit FluidSF by name! 
+
+You can cite the specific version of FluidSF with [Zenodo](https://zenodo.org/records/11406185) or use the following citations (replacing your version number):
+
+> Wagner, C., Pearson, B., & Lee, A. (2024). fluidsf (v0.2.0). Zenodo. https://doi.org/10.5281/zenodo.11406185
+
+
+### Contributing/Developing Guidelines
+
[…]
