### 2025-06-11 af22be7
reformatting and added section for examples
