+++ README.md
@@ -4,7 +4,7 @@
 [![CI](https://github.com/cassidymwagner/FluidSF/actions/workflows/ci.yml/badge.svg?branch=main)](https://github.com/cassidymwagner/FluidSF/actions/workflows/ci.yml)
 
 
-### Overview
+## Overview
 FluidSF is a Python package for calculating structure functions from fluid data. These structure functions can be used to estimate turbulence cascade rates without the constraints of spectral methods. This package serves as a useful tool for analyzing turbulent dynamics in the ocean, atmosphere, and beyond.
 
 <p align="center">
@@ -15,7 +15,7 @@ FluidSF is a Python package for calculating structure functions from fluid data.
 
 **For detailed documentation and examples, see the [FluidSF website](https://cassidymwagner.github.io/fluidsf).**
 
-### Installation
+## Installation
 The easiest method to install FluidSF is with [pip](https://pip.pypa.io/):
 
     $ pip install fluidsf
@@ -28,7 +28,7 @@ You can also fork/clone this repository to your local machine and install it loc
 
 
 
-### Quickstart
+## Quickstart
 
 Once FluidSF is installed, you can load the module into Python and run some basic calculations with any data. Here we'll initialize linearly increasing velocity fields. For more detail on this example, [see the full notebook on the FluidSF website.](https://cassidymwagner.github.io/fluidsf/qs.html)
 
@@ -67,14 +67,14 @@ plt.show()
 <img src="https://github.com/cassidymwagner/fluidsf/blob/main/docs/quickstart.png" alt="Advective structure function plots" width="400"/>
 </p>
 
-### "Can I use FluidSF with my data?"
+## "Can I use FluidSF with my data?"
 
 Hopefully! FluidSF was initially developed for numerical simulations and satellite data, but there are of course many different types of data. If you are interested in using this package but you are unsure how to use it with your dataset, please reach out and we are happy to assist! 
 
 The best way to communicate about your data needs is to [start a discussion](https://github.com/cassidymwagner/fluidsf/discussions) where you can describe your dataset and what you're hoping to learn with FluidSF. Before starting a discussion you can check through other discussion posts or review the open (and closed) issues to see if any other users have a similar question or dataset. 
 
 We have plans to support many different types of data, especially oceanographic data, but we encourage any users to engage with us so we can make sure we support as many datasets as possible!
-### Citing
+## Citing
 
 If you use FluidSF in your research or educational activities, we would be grateful if you credit FluidSF by name! 
 
@@ -83,7 +83,7 @@ You can cite the specific version of FluidSF with [Zenodo](https://zenodo.org/re
 > Wagner, C., Pearson, B., & Lee, A. (2024). fluidsf (v0.2.0). Zenodo. https://doi.org/10.5281/zenodo.11406185
 
 
-### Contributing/Developing Guidelines
+## Contributing/Developing Guidelines
 
 This project welcomes contributions and suggestions. The following contribution/development guidelines have been modified from those provided with [pyqg](https://github.com/pyqg/pyqg), an open-source software package for simulating geophysical fluids. Feel free to open an issue, submit a pull request, and/or [contact the owner directly](https://github.com/cassidymwagner).
 
@@ -126,7 +126,7 @@ After your pull request is merged, you can switch back to the main branch, rebas
     $ git rebase upstream/main
     $ git branch -d cool_new_feature
 
-### Virtual Testing Environment
+### Virtual testing environment
 
 This is how to create a virtual environment into which to test-install FluidSF, install it, check the version, and tear down the virtual environment.
 
@@ -146,7 +146,7 @@ Assuming a local version of FluidSF, install optional dependencies with pip:
     $ pip install .'[examples]'
     $ pip install .'[test]'
 
-### Adding Tests for New Functionality
[…]
