+++ docs/source/index.rst
@@ -36,30 +36,132 @@ You can also fork/clone the `FluidSF repository <https://github.com/cassidymwagn
 
    pip install .
 
+**List of optional dependencies to run example notebooks**: `matplotlib <https://matplotlib.org>`_, `seaborn <https://seaborn.pydata.org>`_, `h5py <https://www.h5py.org>`_, `scipy <https://scipy.org>`_, `xarray <https://xarray.dev>`_
+
 
 .. _Citing:
 
 Citing
 ******
 
-If you use FluidSF in your research or educational activities, we would appreciate it if you cite this work.
+If you use FluidSF in your research or educational activities, we would be grateful if you credit FluidSF by name! 
 
-.. .. code-block:: bibtex
+You can cite the specific version of FluidSF with `Zenodo <https://zenodo.org/records/11406185>`_` or use the following citations (replacing your version number): ::
 
-..    WIP
+   Wagner, C., Pearson, B., & Lee, A. (2024). fluidsf (v0.2.0). Zenodo. https://doi.org/10.5281/zenodo.11406185
 
-.. _Development and Contributing:
 
-Development and Contributing
-****************************
+.. _Contributing/Developing Guidelines:
 
-The development of FluidSF started as a Ph.D. project for Cassidy Wagner, 
-advised by Brodie Pearson. This software package is based upon work supported 
-by the National Science Foundation under Grant No. 2023721.
+Contributing/Developing Guidelines
+**********************************
 
-The developers welcome contributions and suggestions 
-through the `FluidSF GitHub repository <https://github.com/cassidymwagner/FluidSF>`_
-Feel free to open an issue, submit a pull request, and/or 
-`contact the owner directly <https://github.com/cassidymwagner>`_.
+This project welcomes contributions and suggestions. The following contribution/development guidelines have been modified from those provided with `pyqg <https://github.com/pyqg/pyqg>`_, an open-source software package for simulating geophysical fluids. Feel free to open an issue, submit a pull request, and/or `contact the owner directly <https://github.com/cassidymwagner>`_.
 
-Any opinions, findings, and conclusions or recommendations expressed in this package are those of the authors and do not necessarily reflect the views of the National Science Foundation.Any opinions, findings, and conclusions or recommendations expressed in this package are those of the authors and do not necessarily reflect the views of the National Science Foundation.
+Anyone interested in helping to develop FluidSF needs to create their own fork of our `git repository <https://github.com/cassidymwagner/fluidsf>`_. (Follow the github `forking instructions <https://help.github.com/articles/fork-a-repo/>`_. You will need a github account.)
+
+Clone your fork on your local machine. ::
+
+    $ git clone git@github.com:USERNAME/fluidsf
+
+(In the above, replace USERNAME with your github user name.)
+
+Then set your fork to track the upstream FluidSF repo. ::
+
+    $ cd fluidsf
+    $ git remote add upstream git://github.com/fluidsf/fluidsf.git
+
+You will want to periodically sync your main branch with the upstream main. ::
+
+    $ git fetch upstream
+    $ git rebase upstream/main
+
+Never make any commits on your local main branch. Instead open a feature branch for every new development task. ::
+
+    $ git checkout -b cool_new_feature
+
+(Replace `cool_new_feature` with an appropriate description of your feature.) At this point you work on your new feature, using `git add` to add your changes. When your feature is complete and well tested, commit your changes ::
+
+    $ git commit -m 'did a bunch of great work'
+
+and push your branch to github. ::
+
+    $ git push origin cool_new_feature
+
[…]
