### 2025-06-11 dc54cd8
syntax typos and reorganized headers
