+++ docs/source/index.rst
@@ -24,17 +24,13 @@ This package can be applied to data from general fluid flows, although our resea
 Installing
 **********
 
-The easiest method to install FluidSF is with `pip <https://pip.pypa.io/>`_:
+The easiest method to install FluidSF is with `pip <https://pip.pypa.io/>`_: ::
 
-.. code-block:: bash
+   $ pip install fluidsf
 
-   pip install fluidsf
+You can also fork/clone the `FluidSF repository <https://github.com/cassidymwagner/FluidSF>`_ to your local machine and install it locally with pip as well: ::
 
-You can also fork/clone the `FluidSF repository <https://github.com/cassidymwagner/FluidSF>`_ to your local machine and install it locally with pip as well:
-
-.. code-block:: bash
-
-   pip install .
+   $ pip install .
 
 **List of optional dependencies to run example notebooks**: `matplotlib <https://matplotlib.org>`_, `seaborn <https://seaborn.pydata.org>`_, `h5py <https://www.h5py.org>`_, `scipy <https://scipy.org>`_, `xarray <https://xarray.dev>`_
 
@@ -46,7 +42,7 @@ Citing
 
 If you use FluidSF in your research or educational activities, we would be grateful if you credit FluidSF by name! 
 
-You can cite the specific version of FluidSF with `Zenodo <https://zenodo.org/records/11406185>`_` or use the following citations (replacing your version number): ::
+You can cite the specific version of FluidSF with `Zenodo <https://zenodo.org/records/11406185>`_ or use the following citations (replacing your version number): 
 
    Wagner, C., Pearson, B., & Lee, A. (2024). fluidsf (v0.2.0). Zenodo. https://doi.org/10.5281/zenodo.11406185
 
@@ -62,68 +58,71 @@ Anyone interested in helping to develop FluidSF needs to create their own fork o
 
 Clone your fork on your local machine. ::
 
-    $ git clone git@github.com:USERNAME/fluidsf
+   $ git clone git@github.com:USERNAME/fluidsf
 
 (In the above, replace USERNAME with your github user name.)
 
 Then set your fork to track the upstream FluidSF repo. ::
 
-    $ cd fluidsf
-    $ git remote add upstream git://github.com/fluidsf/fluidsf.git
+   $ cd fluidsf
+   $ git remote add upstream git://github.com/fluidsf/fluidsf.git
 
 You will want to periodically sync your main branch with the upstream main. ::
 
-    $ git fetch upstream
-    $ git rebase upstream/main
+   $ git fetch upstream
+   $ git rebase upstream/main
 
 Never make any commits on your local main branch. Instead open a feature branch for every new development task. ::
 
-    $ git checkout -b cool_new_feature
+   $ git checkout -b cool_new_feature
 
-(Replace `cool_new_feature` with an appropriate description of your feature.) At this point you work on your new feature, using `git add` to add your changes. When your feature is complete and well tested, commit your changes ::
+(Replace ``cool_new_feature`` with an appropriate description of your feature.) At this point you work on your new feature, using ``git add`` to add your changes. When your feature is complete and well tested, commit your changes ::
 
-    $ git commit -m 'did a bunch of great work'
+   $ git commit -m 'did a bunch of great work'
 
 and push your branch to github. ::
 
-    $ git push origin cool_new_feature
+   $ git push origin cool_new_feature
 
 At this point, you go find your fork on github.com and create a `pull request <https://help.github.com/articles/using-pull-requests/>`_. Clearly describe what you have done in the comments. If your pull request fixes an issue or adds a useful new feature, the team will gladly merge it.
 
 After your pull request is merged, you can switch back to the main branch, rebase, and delete your feature branch. You will find your new feature incorporated into fluidsf. ::
 
-    $ git checkout main
-    $ git fetch upstream
[…]
