### 2025-06-11 a1649d6
removed warnings from pip install
