+++ .github/workflows/sphinx.yml
@@ -36,7 +36,6 @@ jobs:
         pip install sphinx-rtd-theme
         pip install sphinx-press-theme
         pip install furo
-        pip install warnings
 
     - uses: actions/checkout@master
     - uses: nikeee/setup-pandoc@v1
