+++ _sources/index.rst.txt
@@ -24,17 +24,15 @@ This package can be applied to data from general fluid flows, although our resea
 Installing
 **********
 
-The easiest method to install FluidSF is with `pip <https://pip.pypa.io/>`_:
+The easiest method to install FluidSF is with `pip <https://pip.pypa.io/>`_: ::
 
-.. code-block:: bash
+   $ pip install fluidsf
 
-   pip install fluidsf
+You can also fork/clone the `FluidSF repository <https://github.com/cassidymwagner/FluidSF>`_ to your local machine and install it locally with pip as well: ::
 
-You can also fork/clone the `FluidSF repository <https://github.com/cassidymwagner/FluidSF>`_ to your local machine and install it locally with pip as well:
+   $ pip install .
 
-.. code-block:: bash
-
-   pip install .
+**List of optional dependencies to run example notebooks**: `matplotlib <https://matplotlib.org>`_, `seaborn <https://seaborn.pydata.org>`_, `h5py <https://www.h5py.org>`_, `scipy <https://scipy.org>`_, `xarray <https://xarray.dev>`_
 
 
 .. _Citing:
@@ -42,19 +40,128 @@ You can also fork/clone the `FluidSF repository <https://github.com/cassidymwagn
 Citing
 ******
 
-If you use FluidSF in your research or educational activities, we would appreciate it if you cite this work.
+If you use FluidSF in your research or educational activities, we would be grateful if you credit FluidSF by name! 
 
-.. .. code-block:: bibtex
+You can cite the specific version of FluidSF with `Zenodo <https://zenodo.org/records/11406185>`_ or use the following citations (replacing your version number): 
 
-..    WIP
+   Wagner, C., Pearson, B., & Lee, A. (2024). fluidsf (v0.2.0). Zenodo. https://doi.org/10.5281/zenodo.11406185
 
-.. _Development and Contributing:
 
-Development and Contributing
-****************************
+.. _Contributing/Developing Guidelines: 
 
-The development of FluidSF started as a Ph.D. project for Cassidy Wagner, 
-advised by Brodie Pearson. The developers welcome contributions and suggestions 
-through the `FluidSF GitHub repository <https://github.com/cassidymwagner/FluidSF>`_
-Feel free to open an issue, submit a pull request, and/or 
-`contact the owner directly <https://github.com/cassidymwagner>`_.
+Contributing/Developing Guidelines
+**********************************
+
+This project welcomes contributions and suggestions. The following contribution/development guidelines have been modified from those provided with `pyqg <https://github.com/pyqg/pyqg>`_, an open-source software package for simulating geophysical fluids. Feel free to open an issue, submit a pull request, and/or `contact the owner directly <https://github.com/cassidymwagner>`_.
+
+Anyone interested in helping to develop FluidSF needs to create their own fork of our `git repository <https://github.com/cassidymwagner/fluidsf>`_. (Follow the github `forking instructions <https://help.github.com/articles/fork-a-repo/>`_. You will need a github account.)
+
+Clone your fork on your local machine. ::
+
+   $ git clone git@github.com:USERNAME/fluidsf
+
+(In the above, replace USERNAME with your github user name.)
+
+Then set your fork to track the upstream FluidSF repo. ::
+
+   $ cd fluidsf
+   $ git remote add upstream git://github.com/fluidsf/fluidsf.git
+
+You will want to periodically sync your main branch with the upstream main. ::
+
+   $ git fetch upstream
+   $ git rebase upstream/main
+
+Never make any commits on your local main branch. Instead open a feature branch for every new development task. ::
+
+   $ git checkout -b cool_new_feature
+
+(Replace ``cool_new_feature`` with an appropriate description of your feature.) At this point you work on your new feature, using ``git add`` to add your changes. When your feature is complete and well tested, commit your changes ::
+
[…]
