### 2025-06-11 cassidymwagner/fluidsf#36 @cassidymwagner
We have made all of these changes so far except for submitting to conda-forge, which we have been leaving for the last step once all remaining issues have been resolved.

We have added the documentation changes into the repository's README as well as the homepage of the website (PRs #55 and #57).

We have added `test` and `examples` options in `pyproject.toml` for contributors to install these dependencies easily (PR #45), with instructions in the README/website.
