### 2025-06-11 a400efa
removed myst-nb
