+++ docs/source/conf.py
@@ -43,7 +43,6 @@ extensions = [
     "sphinx.ext.intersphinx",
     "numpydoc",
     "nbsphinx",
-    "myst_nb",
 ]
 templates_path = ["_templates"]
 exclude_patterns = ["tests/test_*", "test_*"]
