+++ docs/source/conf.py
@@ -66,3 +66,4 @@ nbsphinx_execute_arguments = [
     "--InlineBackend.figure_formats={'png'}",
     "--InlineBackend.rc={'figure.dpi': 200}",
 ]
+
