### 2025-06-12 4d45391
added nbstripout to remove notebook output before commits
