+++ .pre-commit-config.yaml
@@ -15,3 +15,8 @@ repos:
     rev: v1.5.1
     hooks:
       - id: mypy
+
+  - repo: https://github.com/kynan/nbstripout
+    rev: 0.8.1
+    hooks:
+      - id: nbstripout
