### 2025-06-12 1fe2da7
Revert "duplicated linting CI that already exists"
This reverts commit 6ab54359c89fd71bf8398e30a45401caf5b153d9.
