### 2025-06-12 ef2f4b0
specify only examples notebooks
