+++ .pre-commit-config.yaml
@@ -20,3 +20,4 @@ repos:
     rev: 0.8.1
     hooks:
       - id: nbstripout
+        files: "^examples/.*/.*\\.ipynb$"
