### 2025-06-12 f1fe6a5
add PR guidelines to website docs
