### 2025-06-12 7d683cb
disabling paper compilation unless commit is tagged with "paper*"
