+++ docs/source/index.rst
@@ -107,8 +107,8 @@ The FluidSF team is happy to help with the pull request process! Here are a few 
 4. GitHub Actions will run several processes to make sure your PR won't break FluidSF or change code formatting:
 
    - In your local repository, use `pre-commit <https://pre-commit.com>`_ and `pytest <https://docs.pytest.org/en/stable/>`_ to make sure your changes pass FluidSF's linting check and test suite.
-   - After you install `pre-commit` you can run ``pre-commit run --all-files`` (or see `pre-commit` documentation to set it up to run automatically with ``git commit``).
-   - See :ref:`Adding tests for new functionality` for details on running `pytest`.
+   - After you install ``pre-commit`` you can run ``pre-commit run --all-files`` (or see ``pre-commit`` documentation to set it up to run automatically with ``git commit``).
+   - See :ref:`Adding tests for new functionality` for details on running ``pytest``.
    - Once your local code passes these checks, push your changes to remote and you should see that your PR passes these checks as well! 🎉
 
 Please don't hesitate to reach out to the FluidSF team if you are having trouble -- we are grateful for your contributions!!
