### 2025-06-12 cassidymwagner/fluidsf#36 @cassidymwagner
@MaceKuailv would it be sufficient to submit to conda-forge once we receive final approval from the reviewers/editor? After this happens we plan to deploy to PyPI with the current FluidSF version, which will streamline the process of [building a recipe for conda-forge](https://conda-forge.org/docs/maintainer/adding_pkgs/#generating-the-recipe) (using `grayskull`).
