### 2025-08-11 cassidymwagner/fluidsf#37 @MaceKuailv
splendid! I have tried all the notebook using pooch. They all work and run fast. good work!
