### 2025-08-11 #7873 @MaceKuailv
I have checked everything again, and I am happy with the state of the package. Make sure to add the package to conda-forge. My final suggestion is to add a yml file in the github repo for developers, although that is just for convenience.
