### 2025-08-14 948dd48
created environment.yaml file for contributors
