+++ environment.yaml
@@ -1,0 +1,15 @@
+name: fluidsf-dev
+
+dependencies:
+  - python>=3.7
+  - numpy>=1,<3
+  - pandas>=2.1.0
+  - geopy>=2
+  - pytest>=7,<8
+  - matplotlib>=3.7.0
+  - seaborn>=0.12.0
+  - h5py>=3.10.0
+  - scipy>=1.11.0
+  - xarray>=2024.1.0
+  - pooch>=1.8.0
+  - h5netcdf>=1.3.0
