### 2025-09-29 #7873 @kyleniemeyer
Hi @cassidymwagner, just a few minor comments on the paper itself:
- Can you add a bit more detail to the author affiliation(s)? Normally department/unit would go here in addition to university.
- Throughout, could you spell out numbers less than 10 (i.e., "two" instead of "2").
