### 2025-10-03 #7873 @cassidymwagner
Thank you for the comments @kyleniemeyer! Would you mind checking [this PR](https://github.com/cassidymwagner/fluidsf/pull/67) to make sure we've incorporated your comments sufficiently? I'll update the package version on the repo/Zenodo/PyPI afterwards.
