### 2025-10-03 #7873 @kyleniemeyer
@cassidymwagner for the affiliation, I think this would be sufficient:
> College of Earth, Ocean, and Atmospheric Sciences, Oregon State University, Corvallis, OR, USA

Otherwise, looks good!
