### 2025-10-03 #7873 @kyleniemeyer
Since only the paper is changing, there is no need to update the archive - that's just for the software itself. JOSS archives the paper
